pyproject.toml
[project]
name = "fil-rouge-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncmy>=0.2.11",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "email-validator>=2.3.0",
    "fastapi>=0.133.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
